Start of candle attempt, checkpoint
candles replaces only torch, doesn't provide ways to get or use
pretrained models. Does seem working and usable though so far as a
replacement to torch but translation is hard not really knowing the code
on either side and it's young so documentation is light.

Going to try `rust-bert`, which uses `tch` (by the same author as
`candles`) as a backend. Might be possible to use `rust-bert` with
`candles`.